Python/Django/SchoolWebsite/school_site/requirements.txt
Django==2.2.2
django-crispy-forms==1.7.2
django-filebrowser-no-grappelli==3.7.8
django-fluent-contents==2.0.5
django-fluent-utils==2.0.1
django-parler==1.9.2
django-polymorphic==2.0.3
django-tag-parser==3.1
django-template-analyzer==1.6.1
django-tinymce4-lite==1.7.5
django-wysiwyg==0.8.0
future==0.17.1
html5lib==1.0.1
jsmin==2.2.2
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
transliterate==1.10.2
translitua==1.2.4
webencodings==0.5.1
